+++ src/com/intellij/idea/plugin/hybris/impex/psi/impl/ImpexParameterMixin.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -26,6 +26,7 @@ import com.intellij.idea.plugin.hybris.impex.psi.references.ImpexDocumentIdRefer
 import com.intellij.idea.plugin.hybris.impex.psi.references.ImpexFunctionTSAttributeReference
 import com.intellij.idea.plugin.hybris.impex.psi.references.ImpexFunctionTSItemReference
 import com.intellij.lang.ASTNode
+import com.intellij.openapi.util.removeUserData
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiReference
 import com.intellij.psi.PsiReferenceBase
@@ -74,8 +75,8 @@ abstract class ImpexParameterMixin(astNode: ASTNode) : ASTWrapperPsiElement(astN
     }
 
     override fun subtreeChanged() {
-        putUserData(ImpexFunctionTSItemReference.CACHE_KEY, null)
-        putUserData(ImpexFunctionTSAttributeReference.CACHE_KEY, null)
+        removeUserData(ImpexFunctionTSItemReference.CACHE_KEY)
+        removeUserData(ImpexFunctionTSAttributeReference.CACHE_KEY)
     }
 
     companion object {