+++ src/com/intellij/idea/plugin/hybris/system/type/psi/reference/result/EnumResolveResult.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -18,13 +18,12 @@
 
 package com.intellij.idea.plugin.hybris.system.type.psi.reference.result
 
-import com.intellij.idea.plugin.hybris.psi.reference.TSReferenceBase
 import com.intellij.idea.plugin.hybris.system.type.meta.model.TSMetaEnum
 import com.intellij.idea.plugin.hybris.system.type.model.EnumType
 
 class EnumResolveResult(
     val meta: TSMetaEnum
-) : TSReferenceBase.TSResolveResult {
+) : TSResolveResult {
     private val myDom: EnumType? = meta.retrieveDom()
     override fun getElement() = myDom?.code?.xmlAttributeValue
     override fun isValidResult() = (myDom?.isValid ?: false) && element != null