+++ src/com/intellij/idea/plugin/hybris/system/type/psi/reference/result/RelationEndResolveResult.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -18,14 +18,13 @@
 
 package com.intellij.idea.plugin.hybris.system.type.psi.reference.result
 
-import com.intellij.idea.plugin.hybris.psi.reference.TSReferenceBase
 import com.intellij.idea.plugin.hybris.system.type.meta.model.TSMetaRelation
 import com.intellij.idea.plugin.hybris.system.type.model.RelationElement
 import com.intellij.psi.PsiElement
 
 class RelationEndResolveResult(
     val meta: TSMetaRelation.TSMetaRelationElement
-) : TSReferenceBase.TSResolveResult {
+) : TSResolveResult {
     private val myDom: RelationElement? = meta.retrieveDom()
     override fun getElement(): PsiElement? = if (meta.isNavigable) myDom
         ?.qualifier